fix: cast $1 to bigint in the stale run recovery update
Postgres deduces $1 as bigint from "last_failure_epoch = $1" and as integer
from "$1 + 3600", rejects the statement with 42P08 inconsistent types deduced
for parameter $1, and the worker exits before it can start. Cast both
references explicitly.

Verified with PREPARE against the production database: the uncast form fails
with 42P08, the cast form prepares with parameter_types {bigint}.

The existing test drives recoverPostgresStaleRuns through a stub pool that
accepts any SQL, so it could not catch a parameter type error. It now asserts
that every $1 reference carries the cast.

+++ server/ingestion/workerStore.js
@@ -590,8 +590,8 @@ async function recoverPostgresStaleRuns(pool) {
       UPDATE company_sync_state st
       SET
         consecutive_failures = COALESCE(st.consecutive_failures, 0) + 1,
-        last_failure_epoch = $1,
-        next_sync_epoch = GREATEST(COALESCE(st.next_sync_epoch, 0), $1 + 3600),
+        last_failure_epoch = $1::bigint,
+        next_sync_epoch = GREATEST(COALESCE(st.next_sync_epoch, 0), $1::bigint + 3600),
         last_error = 'Worker restarted while processing this target',
         updated_at = now()
       FROM ingestion_runs r
